// src/types/generated/CircuitEvidenceView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CircuitCheckpoint } from "./CircuitCheckpoint";
import type { CircuitHistoryEntry } from "./CircuitHistoryEntry";
import type { CircuitStepObservationCoverage } from "./CircuitStepObservationCoverage";

export type CircuitEvidenceView = { entries: Array<CircuitHistoryEntry>, checkpoints: Array<CircuitCheckpoint>, coverage: Array<CircuitStepObservationCoverage>, };